+++ Assets/Survaivor/Character/Combat/Application/DamageFieldSpawnRequest.cs
@@ -22,7 +22,7 @@ public sealed class DamageFieldSpawnRequest : WeaponFireRequest
     /// <param name="followTarget">追従対象</param>
     /// <param name="sourcePow">攻撃元の攻撃力</param>
     /// <param name="areaScale">適用するエリア倍率</param>
-    public DamageFieldSpawnRequest(Vector3 origin, Transform followTarget, int sourcePow, float areaScale) : base(origin, sourcePow)
+    public DamageFieldSpawnRequest(Vector3 origin, Transform followTarget, float sourcePow, float areaScale) : base(origin, sourcePow)
     {
         FollowTarget = followTarget;
         AreaScale = areaScale;
